test/web-platform-tests/to-upstream/FileAPI/file/File-constructor.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>File constructor</title>
<link rel=help href="http://dev.w3.org/2006/webapi/FileAPI/#dfn-file">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
"use strict";

// async testing
async_test(t => {
  const curTime = Date.now();
  const file = new File([], "dummy");
  setTimeout(() => {
    assert_true(Math.abs(curTime - file.lastModified) < 200,
        "lastModified should be generated at File construction if not provided");
    t.done();
  }, 400);
});

</script>
